Refactor connection setup into lib/schema.js
